FURIAX-Xmodule/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
backports-datetime-fromisoformat==2.0.3
beautifulsoup4==4.13.4
certifi==2025.1.31
click==8.1.8
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.12
filetype==1.2.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
Js2Py_3.13==0.74.1
lxml==5.4.0
m3u8==6.0.0
pydantic==2.11.3
pydantic_core==2.33.1
pyjsparser==2.7.1
pymongo==4.12.0
pyotp==2.9.0
python-dotenv==1.1.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.7
starlette==0.46.2
twikit==2.3.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
uvicorn==0.34.2
webvtt-py==0.5.1
